% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit_sel.R
\name{Lambda2}
\alias{Lambda2}
\title{Lambda2 Function - Bivariate Normal Copula}
\usage{
Lambda2(u, v, rho)
}
\arguments{
\item{u}{First probability (must be in (0,1))}

\item{v}{Second probability (must be in (0,1))}

\item{rho}{Correlation parameter (must be in (-1,1))}
}
\value{
Joint probability from bivariate normal copula
}
\description{
Computes the bivariate normal copula joint probability
}
