% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit_sel.R
\name{multinomial_probs}
\alias{multinomial_probs}
\title{Multinomial Probabilities}
\usage{
multinomial_probs(X, beta_list)
}
\arguments{
\item{X}{Matrix of covariates}

\item{beta_list}{List of coefficient vectors for each category (except baseline)}
}
\value{
Matrix of probabilities
}
\description{
Computes multinomial logit probabilities
}
