% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit_sel.R
\name{summary.bootstrap_inference}
\alias{summary.bootstrap_inference}
\title{Summary Method for Bootstrap Inference}
\usage{
\method{summary}{bootstrap_inference}(bootstrap_result)
}
\arguments{
\item{bootstrap_result}{Output from bootstrap_inference function}
}
\description{
Print a summary of bootstrap inference results
}
