% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit_sel.R
\name{Lambda2}
\alias{Lambda2}
\title{Lambda2 Function for Copula}
\usage{
Lambda2(u, v, omega)
}
\arguments{
\item{u}{First probability}

\item{v}{Second probability}

\item{omega}{Dependence parameter}
}
\value{
Joint probability
}
\description{
Computes the bivariate copula joint probability
}
