% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit_sel.R
\name{summary.bootstrap_inference}
\alias{summary.bootstrap_inference}
\title{Summary Method for Bootstrap Inference}
\usage{
\method{summary}{bootstrap_inference}(object, ...)
}
\arguments{
\item{object}{Output from bootstrap_inference function}

\item{...}{Additional arguments (not used)}
}
\description{
Print a summary of bootstrap inference results
}
